Seperated all enum classes to enums file
