+++ utils/sort_stream.py
@@ -1,32 +1,12 @@
 from datetime import date
-from enum import Enum
 from typing import Optional
 from typing import Union
 
 from utils.create_stream import StreamMap
-
-
-class Duration(str, Enum):
-    LONGER = "duration_longer_than"
-    SHORTER = "duration_shorter_than"
-    BETWEEN = "duration_between"
-
-
-class Size(str, Enum):
-    LARGER = "size_larger_than"
-    SMALLER = "size_smaller_than"
-    BETWEEN = "size_between"
-
-
-class Creation(str, Enum):
-    YEAR = "year"
-    MONTH = "month"
-    DAY = "day"
-
-
-class Audio(str, Enum):
-    AUDIO = "audio"
-    NO_AUDIO = "no_audio"
+from utils.enums import Audio
+from utils.enums import Creation
+from utils.enums import Duration
+from utils.enums import Size
 
 
 def _sort_greater_than(
